// src/app/models/cocktail.model.ts
export class Cocktail {
    drinks: Drink[];
}

export class Drink {
    idDrink?: string;
    strDrink?: string;
    strCategory?: string;
    strAlcoholic?: string;
    strGlass?: string;
    strInstructions?: string;
    strDrinkThumb?: string;
    strIngredient1?: string;
    strIngredient2?: string;
    strIngredient3?: string;
}